more fixes

+++ src/lib/awsElasticSearch/index.js
@@ -109,7 +109,7 @@ const addIndexDocument = async (job) => {
       body: document,
       refresh: true, // triggers manual refresh.
     });
-    logger.info(`AWS OpenSearch: Successfully added document to index ${indexName}`);
+    logger.info(`AWS OpenSearch: Successfully added document ${id} to index ${indexName}`);
 
     return { data: job.data, status: res.statusCode, res };
   } catch (error) {
@@ -197,10 +197,11 @@ const updateIndexDocument = async (job) => {
       index: indexName,
       id,
       body,
+      doc_as_upsert: true, // create if doesn't exist.
       refresh: true, // triggers manual refresh.
     });
 
-    logger.info(`AWS OpenSearch: Successfully updated document  index ${indexName} for id ${id}`);
+    logger.info(`AWS OpenSearch: Successfully updated document index ${indexName} for id ${id}`);
     return { data: job.data, status: res.statusCode, res };
   } catch (error) {
     auditLogger.error(`AWS OpenSearch Error: Unable to update the index '${indexName} for id ${id}': ${error.message}`);
@@ -220,14 +221,34 @@ const deleteIndexDocument = async (job) => {
     // Initialize the client.
     const client = passedClient || await getClient();
 
-    // Delete index document.
-    res = await client.delete({
-      index: indexName,
-      id,
-      refresh: true, // triggers manual refresh.
-    });
+    // Check document exists.
+    const query = {
+      query: {
+        match: {
+          id: {
+            query: id,
+          },
+        },
+      },
+    };
 
-    logger.info(`AWS OpenSearch: Successfully deleted document '${id}' for index '${indexName}'`);
+    res = await client.search({
+      index: indexName,
+      body: query,
+    });
+    if (res.body.hits.hits.length >= 1) {
+      // Delete index document.
+      res = await client.delete({
+        index: indexName,
+        id,
+        refresh: true, // triggers manual refresh.
+        body: { ignore_unavailable: true },
+      });
+      logger.info(`AWS OpenSearch: Successfully deleted document '${id}' for index '${indexName}'`);
+    } else {
+      logger.warn(`AWS OpenSearch: Unable to locate a document with '${id}' for index to delete'${indexName}'`);
+    }
+
     return { data: job.data, status: res.statusCode, res };
   } catch (error) {
     auditLogger.error(`AWS OpenSearch Error: Unable to delete document '${id}' for index '${indexName}': ${error.message}`);
@@ -244,6 +265,7 @@ const deleteIndex = async (indexName, passedClient) => {
     // Initialize the client.
     const client = passedClient || await getClient();
 
+    // Delete.
     res = await client.indices.delete({
       index: indexName,
     });
